Split to controller, logic and model
