+++ api/controller.py
@@ -1,27 +1,7 @@
 from flask import Flask, jsonify, abort, make_response, request
+from logic import *
 
 app = Flask(__name__)
-
-books = [
-    {
-        'id': 1,
-        'name': 'Boek 1',
-        'price': 10,
-        'isbn': 11
-    },
-    {
-        'id': 2,
-        'name': 'Boek 2',
-        'price': 20.0,
-        'isbn': 22
-    },
-    {
-        'id': 3,
-        'name': 'Boek 3',
-        'price': 30.0,
-        'isbn': 33
-    }
-]
 
 @app.errorhandler(404)
 def notFoundError(error):
@@ -103,53 +83,6 @@ def deleteBook(id):
         abort(404)
     deleteBookLogic(id)
     return make_response("", 200)
-
-#
-# Logic
-#
-def getAllBookslogic():
-    return books
-
-def getBookByIdLogic(id):
-    book = [book for book in books if book['id'] == id]
-    return book
-
-def addBookLogic(newBookData):
-
-    newBook = {
-        'id': books[-1]['id'] + 1, 
-        'name': newBookData.get('name', ''),
-        'price': newBookData.get('price', 0),
-        'isbn': newBookData.get('isbn', 0),
-    }
-
-    books.append(newBook)
-    return newBook
-
-def editBookLogic(id, updatedBook):
-    index = 0
-    for book in books:
-        if book['id'] == id:
-            break
-        index += 1
-
-    if 'name' in updatedBook:
-        books[index]['name'] = updatedBook['name']
-    if 'isbn' in updatedBook:
-        books[index]['isbn'] = updatedBook['isbn']
-    if 'price' in updatedBook:
-        books[index]['price'] = updatedBook['price']
-
-    return True    
-
-def deleteBookLogic(id):
-    index = 0
-    for book in books:
-        if book['id'] == id:
-            break
-        index += 1
-    books.remove(books[index])
-    return True
 
 if __name__ == '__main__':
     app.run(port=5000, debug=True)  # auto-reload
