+++ api/model.py
@@ -1,0 +1,20 @@
+books = [
+    {
+        'id': 1,
+        'name': 'Boek 1',
+        'price': 10,
+        'isbn': 11
+    },
+    {
+        'id': 2,
+        'name': 'Boek 2',
+        'price': 20.0,
+        'isbn': 22
+    },
+    {
+        'id': 3,
+        'name': 'Boek 3',
+        'price': 30.0,
+        'isbn': 33
+    }
+]
